+++ R/focei.R
@@ -1376,11 +1376,17 @@ attr(rxUiGet.foceiOptEnv, "desc") <- "Get focei optimization environment"
       toupper(x)
     }
   }, character(1))
-  requiredCols <- c("ID", "DV", "TIME", .covNames)
   if (is.null(data$ID)) data$ID <- 1L
-  checkmate::assert_names(names(data), must.include = requiredCols)
+  colnames(data) <- vapply(names(data), function(x) {
+    if (any(x == .covNames)) {
+      x
+    } else {
+      toupper(x)
+    }
+  }, character(1))
   if (is.null(data$EVID) && is.null(data$AMT)) data$EVID <- 0
   if (is.null(data$AMT)) data$AMT <- 0
+  checkmate::assert_names(names(data), must.include = c("DV", "TIME"))
   ## Make sure they are all double amounts.
   for (.v in c("DV", "TIME")) {
     data[[.v]] <- as.double(data[[.v]])